+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.1
+
+* Support latest `pkg/shelf`.
+
 ## 1.0.0
 
 * Initial version.